rest_API/requirements.txt
django
psycopg2-binary
djangorestframework
aiohttp
environs